requirements.txt
accelerate==0.25.0
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-property==0.2.2
async-timeout==4.0.3
asyncstdlib==3.12.0
attrs==23.1.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
curl-cffi==0.5.10
diffusers==0.24.0
distro==1.9.0
duckduckgo_search==4.2
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
Flask==3.0.1
frozenlist==1.4.1
fsspec==2023.12.2
g4f==0.2.0.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.1
hyperframe==5.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
lxml==5.1.0
lz4==4.3.3
magic-filter==1.0.12
MarkupSafe==2.1.3
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.2
openai==1.7.2
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
psutil==5.9.7
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyExecJS==1.5.1
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
rawkit==0.6.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.4
urllib3==2.1.0
uvicorn==0.26.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
